requirements.txt
codecov==2.0.9
coverage==4.4.2
coveralls==1.2.0
pycodestyle==2.3.1
lxml==4.1.1
pytest==3.3.0
pytest-cov==2.5.1
pytest-flake8==0.9.1
python-qpid-proton==0.19.0
